requirements.txt
keras==2.10.0
numpy==1.23.4
Pillow==9.3.0
streamlit==1.14.0
tensorflow==2.10.0
